requirements.txt
faiss-cpu
flask
flask-restful
openai
numpy
pickle
os
scikit-learn